requirements.txt
psycopg2-binary
PyPDF2
tqdm
falcon~=3.1.1
requests~=2.31.0
numpy~=1.19.5
dataclasses
pyyaml~=6.0.1
wbtools==3.1.0
#git+https://github.com/WormBase/wbtools.git@develop
scikit-learn
sent2vec-prebuilt
nltk~=3.6.3
transformers~=4.35.2
datasets
optimum
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
gspread>=5.0.0